// types/goods/GoodsSafeCertiHistDataModel.ts
export type PrGoodsSafeCertiHist = {
    goodsNo: string | null | undefined // 상품번호
    safeCertiGbCd: string | null | undefined // 안전인증구분코드(PR026)
    safeCertiNm: string | null | undefined // 안전인증구분명
    safeCertiNo: string | null | undefined // 안전인증번호
    aplyStrDt: string | null | undefined // 적용시작일자
    aplyEndDt: string | null | undefined // 적용종료일자
    safeCertiOrnNm: string | null | undefined  // 안전인증기관명
}